Modifying to make LineUpQueue compatible with Queue interface of Java. Actually it implements the BlockingQueue<QueueMessage> interface.

+++ src/main/java/com/sangupta/lineup/queues/AbstractInternalQueue.java
@@ -118,7 +118,7 @@ public abstract class AbstractInternalQueue implements InternalQueue {
 	 * @see com.sangupta.lineup.queues.InternalQueue#getMessage(int)
 	 */
 	@Override
-	public QueueMessage getMessage(int longPollTime) throws InterruptedException {
+	public QueueMessage getMessage(long longPollTime) throws InterruptedException {
 		QueueMessage qm = this.QUEUE.poll(longPollTime, TimeUnit.SECONDS);
 		if(qm != null) {
 			removeMessage(qm);
